brownie/requirements.txt
fastapi==0.115.11
psycopg2-binary==2.9.10
pydantic==2.10.6
pandas==2.2.3
uvicorn==0.34.0
feedparser==6.0.11
starlette==0.46.1
h11==0.14.0
idna==3.10
click==8.1.8
charset-normalizer==3.4.1
urllib3==2.3.0
certifi==2025.1.31